# hms-project/docs/ambulance-management-openapi.yaml
openapi: 3.0.3
info:
  title: HMS Ambulance Management API
  description: >-
    This OpenAPI specification documents all ambulance management related
    endpoints, including ambulance inventory management, dispatch system,
    real-time location tracking, driver management, service history,
    and billing integration.
  version: 1.0.0
servers:
  - url: https://hms.example.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local dev server
...
paths:
...
  /ambulances/dispatch:
    get:
      tags: [Ambulance Dispatch]
      summary: Get ambulance dispatch requests with optional filtering
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [REQUESTED, ASSIGNED, DISPATCHED, ARRIVED, COMPLETED, CANCELLED]
        - in: query
          name: id
          schema:
            type: string
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: List of dispatch requests with pagination metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchListResponse'
    post:
      tags: [Ambulance Dispatch]
      summary: Create a new ambulance dispatch request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchInput'
      responses:
        '201':
          description: Dispatch request created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispatch'
    put:
      tags: [Ambulance Dispatch]
      summary: Update an ambulance dispatch request status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                status:
                  type: string
                  enum: [ASSIGNED, DISPATCHED, ARRIVED, COMPLETED, CANCELLED]
              required: [id, status]
      responses:
        '200':
          description: Dispatch request updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispatch'
...
components:
  schemas:
    PagedResponse:
      type: object
      properties:
        data:
          type: array
          items: { type: object }
        meta:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
...
    DispatchListResponse:
      allOf:
        - $ref: '#/components/schemas/PagedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Dispatch'
    DispatchInput:
      type: object
      required: [pickupLocation, dropLocation, purpose, requestedBy]
      properties:
        pickupLocation:
          type: string
        dropLocation:
          type: string
        purpose:
          type: string
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, EMERGENCY]
          default: NORMAL
        patientId:
          type: string
        requestedBy:
          type: string
        notes:
          type: string
    Dispatch:
      allOf:
        - $ref: '#/components/schemas/DispatchInput'
        - type: object
          properties:
            id:
              type: string
            ambulanceId:
              type: string
            driverId:
              type: string
            status:
              type: string
              enum: [REQUESTED, ASSIGNED, DISPATCHED, ARRIVED, COMPLETED, CANCELLED]
            requestedAt:
              type: string
              format: date-time
            dispatchedAt:
              type: string
              format: date-time
            arrivedAt:
              type: string
              format: date-time
            completedAt:
              type: string
              format: date-time
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...